Recovered installLinterReekDeps with idleCallback
Looks like the file itself (linter-reek.js) is only loaded when the activation hook is triggered... but it blocks the opening of the file until the package is done calling activate().

This means that all the optimizations we were doing before to reduce the load during Atom's initial startup are still useful for packages that use activationHooks because we don't want to be blocking the UI any more than we must.